message: move `MessageData` types
Moves the `MessageData` types into the `message_data` module.
